+++ PetriNets/PetriNet.cs
@@ -734,769 +734,4 @@ namespace PetriNets
 		}
 	}
 
-
-	/// <summary>
-	/// Класс для визуального представления MPlace
-	/// </summary>
-	[Serializable]
-	public class Circle : IShape, INotArch
-	{
-		/// <summary>
-		/// Шрифт для написания количества токенов в объекте MPlace связаного с этим объектом
-		/// </summary>
-		SerializableFont font = new SerializableFont("Arial", GraphicsUnit.Pixel, 12, FontStyle.Regular);
-
-		/// <summary>
-		/// Дефолтный цвет фигуры
-		/// </summary>
-		private Color DefaultColor = ControlPaint.Light(Color.Red);
-		/// <summary>
-		/// Цвет фигуры при выделении
-		/// </summary>
-		private Color SelectedColor = Color.Green;
-		/// <summary>
-		/// Радиус для рисования токена
-		/// </summary>
-		private int TokenRadius = 3;
-		/// <summary>
-		/// Смещения для рисования токенов внутри объкта Circe
-		/// </summary>
-		private Point[] offsets;
-
-		/// <summary>
-		/// Конструктор на основании Центра, Радиуса и наимнеования
-		/// </summary>
-		/// <param name="_Center">Место где будет центр рисуемого объекта</param>
-		/// <param name="_Radious">Радиус объекта</param>
-		/// <param name="labelText">наименование</param>
-		public Circle(Point _Center, int _Radious, string labelText)
-		{
-			FillColor = ControlPaint.Light(Color.Red);
-			Center = _Center;
-			Radious = _Radious;
-			label = new Label(new Point(_Center.X + 15, _Center.Y + 15), labelText);
-			inLines = new List<Line>();
-			model = new MPlace(labelText);
-			fillOfssets();
-		}
-		/// <summary>
-		/// Соответсвующая визуальному представлению модель MPlace
-		/// </summary>
-		public MPlace model;
-		/// <summary>
-		/// Линии принадлежащих объекту
-		/// </summary>
-		public List<Line> inLines { get; set; }
-		/// <summary>
-		/// Наименование
-		/// </summary>
-		public Label label { get; set; }
-		/// <summary>
-		/// Цвет заполнения фигуры
-		/// </summary>
-		public Color FillColor { get; set; }
-		/// <summary>
-		/// Центр фигуры
-		/// </summary>
-		public Point Center { get; set; }
-		/// <summary>
-		/// Радиус фигуры
-		/// </summary>
-		public int Radious { get; set; }
-		#region select, draw, move
-		/// <summary>
-		/// Построение GraphicsPath для объекта Circle в виде эллипса 
-		/// </summary>
-		/// <returns>полученый GraphicsPath</returns>
-		public GraphicsPath GetPath()
-		{
-			var path = new GraphicsPath();
-			var p = Center;
-			p.Offset(-Radious, -Radious);
-			path.AddEllipse(p.X, p.Y, 2 * Radious, 2 * Radious);
-			return path;
-		}
-		/// <summary>
-		/// Проверка попадания точки в построенный GraphicsPath для объекта Circle
-		/// </summary>
-		/// <param name="p">Точка для которой проверяется попадание</param>
-		/// <returns></returns>
-		public bool HitTest(Point p)
-		{
-			var result = false;
-			using (var path = GetPath())
-				result = path.IsVisible(p);
-			return result || label.HitTest(p);
-		}
-		/// <summary>
-		/// Метод отрисовки объекта
-		/// </summary>
-		/// <param name="g">Место отрисовки</param>
-		public void Draw(Graphics g)
-		{
-			using (var path = GetPath())
-			using (var brush = new SolidBrush(FillColor))
-				g.FillPath(brush, path);
-			label.Draw(g);
-			drawToken(g,(int)model.tokens);
-		}
-		/// <summary>
-		/// Метод вызываемый при событии onMouseMove. 
-		/// Устанавливает новые координаты центра, поэтому при onPaint перемещается
-		/// </summary>
-		/// <param name="d"></param>
-		public void Move(Point d)
-		{
-			Center = new Point(Center.X + d.X, Center.Y + d.Y);
-			label.Move(d);
-			foreach (Line line in inLines)
-			{
-				if (line.startShape == this) { line.ResizeStart(d); }
-				else { line.Resize(d); }
-			}
-		}
-		#endregion
-
-		#region IShape other methods
-		/// <summary>
-		/// Переименовать объект Circle
-		/// </summary>
-		/// <param name="newName">новое наименование</param>
-		public void RenameLabel(string newName)
-		{
-			this.label.Text = newName;
-		}
-		/// <summary>
-		/// Поменять цвет на цвет выделения
-		/// </summary>
-		public void Select()
-		{
-			FillColor = SelectedColor;
-		}
-		/// <summary>
-		/// Поменять цвет на дефолтный
-		/// </summary>
-		public void Unselect()
-		{
-			FillColor = DefaultColor;
-		}
-		/// <summary>
-		/// Получить центр фигуры
-		/// </summary>
-		/// <returns>Центр фигуры</returns>
-		public Point getCenter()
-		{
-			return Center;
-		}
-		/// <summary>
-		/// Получить линии связанные с объектом
-		/// </summary>
-		/// <returns>Список связанных линий</returns>
-		public List<Line> getLines() { return inLines; }
-		/// <summary>
-		/// Удалить фигуру из данного массива объектов типа IShape. А так же удаляет все зависимости из модели
-		/// </summary>
-		/// <param name="shapes">массива объектов типа IShape</param>
-		public bool delete(List<IShape> shapes)
-		{
-			try
-			{
-				shapes.Remove(this);
-				foreach (var line in inLines)
-				{
-					TRectangle rectangle;
-					if (line.startShape is TRectangle)
-					{
-						rectangle = line.startShape as TRectangle;
-						deleteReference(line.mArc, rectangle.model.inPlaces);
-						rectangle.inLines.Remove(line);
-					}
-					else if (line.endShape is TRectangle)
-					{
-						rectangle = line.endShape as TRectangle;
-						deleteReference(line.mArc, rectangle.model.outPlaces);
-						rectangle.inLines.Remove(line);
-					}
-					shapes.Remove(line);
-				}
-				return true;
-			}
-			catch { return false; }
-		}
-		/// <summary>
-		/// Метод удаления зависимостей из данного листа объекто типа MArc
-		/// </summary>
-		/// <param name="arc">арка с объекто MPlace для которого удаляем зависимость</param>
-		/// <param name="arcs">Список откдуа удаляем зависимость</param>
-		/// <returns>Результат удаления</returns>
-		private bool deleteReference(MArc arc, List<MArc> arcs)
-		{
-			return PetriNetsClassLibrary.PetriNet.CTransition.removeLink(arc, arcs);
-		}
-		#endregion
-
-		#region token draw method
-		void drawToken(Graphics g, int numberOfToken)
-		{
-			if (numberOfToken > 5)
-			{
-				var z = Center;
-				z.Offset(-6, -6);
-				g.DrawString(numberOfToken + "", font.ToFont(), Brushes.Black, z);
-				return;
-			}
-			for (int i = 0; i < numberOfToken; i++) {
-				var p = Center;
-				p.Offset(offsets[i]);
-				g.FillEllipse(Brushes.Black, p.X - TokenRadius, p.Y - TokenRadius, TokenRadius * 2, TokenRadius * 2);
-			}
-		}
-
-
-		void fillOfssets()
-		{
-			offsets = new Point[5] {new Point(0, 0),
-									new Point(+Radious - 12, -Radious + 12),
-									new Point(-Radious + 12, +Radious - 12),
-									new Point(Radious - 12, Radious - 12),
-									new Point(-Radious + 12, -Radious + 12)
-		};
-		}
-		#endregion
-
-		/// <summary>
-		/// Метод получения ближайшей точки границы объекта Circle относительно переденаной точки
-		/// </summary>
-		/// <param name="Location">Поиск ближайшей точки относительно этой точки</param>
-		/// <returns>Ближайшая точка на границе</returns>
-		public PointF getClosestEdge(PointF Location)
-		{
-			float sqrt = (float)Math.Sqrt(Math.Pow(Location.X - Center.X, 2) + Math.Pow(Location.Y - Center.Y, 2));
-			float cx = Center.X + Radious * (Location.X - Center.X) / sqrt;
-			float cy = Center.Y + Radious * (Location.Y - Center.Y) / sqrt;
-			return new PointF(cx, cy);
-		}
-
-	}
-
-
-	/// <summary>
-	/// Класс для визуального представления MTransition
-	/// </summary>
-	[Serializable]
-	public class TRectangle : IShape, INotArch
-	{
-		/// <summary>
-		/// Цвет рисования
-		/// </summary>
-		private Color DefaultColor = ControlPaint.Light(Color.DarkBlue);
-		/// <summary>
-		/// Цвет выделения
-		/// </summary>
-		private Color SelectedColor = Color.Green;
-
-		/// <summary>
-		/// Конструктор нового объекта TRectangle с известным местоположением и габаритами
-		/// </summary>
-		/// <param name="_Center">Центр-Позиция рисования</param>
-		/// <param name="_height">Выоста</param>
-		/// <param name="_width">Ширина</param>
-		/// <param name="labelText">наименование</param>
-		public TRectangle(Point _Center, int _height, int _width, string labelText)
-		{
-			FillColor = ControlPaint.Light(Color.DarkBlue);
-			Center = _Center;
-			height = _height;
-			width = _width;
-			label = new Label(new Point(_Center.X + 10, _Center.Y + 10), labelText);
-			inLines = new List<Line>();
-			model = new MTransition(labelText);
-			PetriNetsClassLibrary.PetriNet.CTransition.allTransition.Add(model);
-		}
-		/// <summary>
-		/// Модель MTransition соответствующая этому объекту TRectangle
-		/// </summary>
-		public MTransition model;
-		/// <summary>
-		/// Все зависимые от этого объекта линии
-		/// </summary>
-		public List<Line> inLines { get; set; }
-		/// <summary>
-		/// наименование
-		/// </summary>
-		public Label label { get; set; }
-		/// <summary>
-		/// Выбранный цвет заполнения
-		/// </summary>
-		public Color FillColor { get; set; }
-		/// <summary>
-		/// Центр фигуры
-		/// </summary>
-		public Point Center { get; set; }
-		/// <summary>
-		/// Высота фигуры
-		/// </summary>
-		public int height { get; set; }
-		/// <summary>
-		/// Ширина фигуры
-		/// </summary>
-		public int width { get; set; }
-		/// <summary>
-		/// Получить объект GraphicsPath формы Прямоугольника
-		/// </summary>
-		/// <returns>объект GraphicsPath формы Прямоугольника</returns>
-		public GraphicsPath GetPath()
-		{
-			var path = new GraphicsPath();
-			var p = Center;
-			p.Offset(-width / 2, -height / 2);
-			path.AddRectangle(new Rectangle(p.X, p.Y, width, height));
-			return path;
-		}
-		/// <summary>
-		/// Проверка попадания точки в контур GraphicsPath
-		/// </summary>
-		/// <param name="p">проверяемая точка</param>
-		/// <returns>Результат попадания</returns>
-		public bool HitTest(Point p)
-		{
-			var result = false;
-			using (var path = GetPath())
-				result = path.IsVisible(p);
-			return result || label.HitTest(p);
-		}
-		/// <summary>
-		/// Рисование контура GraphicsPath
-		/// </summary>
-		/// <param name="g">Место для рисования</param>
-		public void Draw(Graphics g)
-		{
-			using (var path = GetPath())
-			using (var brush = new SolidBrush(FillColor))
-				g.FillPath(brush, path);
-			//var bounds = GetPath().GetBounds();
-			//g.DrawLine(Pens.Black, new Point((int)bounds.Left - 10, Center.Y + height / 2), new Point((int)bounds.Left - 10, Center.Y - height / 2));
-			//var pos = getLineEquation(new Point((int)bounds.Left - 10, Center.Y + height / 2), new Point((int)bounds.Left - 10, Center.Y - height / 2), new Point(Center.X, Center.Y));
-			//g.DrawEllipse(Pens.Black, pos.X,pos.Y, 5, 5);
-			label.Draw(g);
-		}
-		/// <summary>
-		/// Метод для имитации передвижения объекта. Замена старого местоположения на новое
-		/// </summary>
-		/// <param name="d">Смещение относительно старой позиции</param>
-		public void Move(Point d)
-		{
-			Center = new Point(Center.X + d.X, Center.Y + d.Y);
-			label.Move(d);
-			foreach (Line line in inLines)
-			{
-				if (line.startShape == this) { line.ResizeStart(d); }
-				else { line.Resize(d); }
-			}
-		}
-		/// <summary>
-		/// Переименовать объект
-		/// </summary>
-		/// <param name="newName"></param>
-		public void RenameLabel(string newName)
-		{
-			this.label.Text = newName;
-		}
-		/// <summary>
-		/// Поменять цвет объекта на SelectedColor
-		/// </summary>
-		public void Select()
-		{
-			FillColor = SelectedColor;
-		}
-		/// <summary>
-		/// Поменять цвет объекта на DefaultColor
-		/// </summary>
-		public void Unselect()
-		{
-			FillColor = DefaultColor;
-		}
-		/// <summary>
-		/// Получить позицию центра фигуры
-		/// </summary>
-		/// <returns>позиция центра фигуры</returns>
-		public Point getCenter() { return Center; }
-
-		/// <summary>
-		/// Получить все зависимые от объекта TRectangle линии
-		/// </summary>
-		/// <returns>все зависимые от объекта линии</returns>
-		public List<Line> getLines() { return inLines; }
-
-		/// <summary>
-		/// Удалить из указаного листа IShape текущий объект типа TRectangle
-		/// </summary>
-		/// <param name="shapes"></param>
-		public bool delete(List<IShape> shapes)
-		{
-			try
-			{
-				foreach (var line in inLines)
-				{
-					Circle circle;
-					if (line.startShape is Circle)
-					{
-						circle = line.startShape as Circle;
-					}
-					else
-					{
-						circle = line.endShape as Circle;
-					}
-					circle.inLines.Remove(line);
-					shapes.Remove(line);
-				}
-				shapes.Remove(this);
-				return true;
-			}
-			catch { return false; }
-		}
-
-		/// <summary>
-		/// Метод выделения объекта Rectangle для режима выполнения сетей петри
-		/// </summary>
-		/// <param name="g"></param>
-		public void selectRectangle(Graphics g)
-		{
-			Pen pen;
-			if (this.model.isEnable)
-				pen = new Pen(Brushes.Green, 10);
-			else pen = new Pen(Brushes.Red, 10);
-			using (var path = GetPath())
-				g.DrawPath(pen, path);
-		}
-
-		/// <summary>
-		/// Метод получения ближайшей точки границы объекта TRectangle относительно переденаной точки
-		/// </summary>
-		/// <param name="Location">Поиск относительно этой точки</param>
-		/// <returns>Ближайшая точка на границе</returns>
-		public PointF getClosestEdge(PointF Location)
-		{
-			var bounds = GetPath().GetBounds();
-			float d1, d2, d3, d4;
-			double l1, l2, l3, l4;
-			var leftClosest = FindClosestPointOnLine(new PointF((int)bounds.Left, Center.Y + height / 2 - 5),
-											  new PointF((int)bounds.Left, Center.Y - height / 2 + 5),
-											  Location, out d1, out l1);
-			var rightClosest = FindClosestPointOnLine(new PointF((int)bounds.Right, Center.Y + height / 2 - 5),
-											  new PointF((int)bounds.Right, Center.Y - height / 2 + 5),
-											  Location, out d2, out l2);
-			var bottomClosest = FindClosestPointOnLine(new PointF(Center.X + width / 2, (int)bounds.Bottom),
-											  new PointF(Center.X - width / 2, (int)bounds.Bottom),
-											  Location, out d3, out l3);
-			var topClosest = FindClosestPointOnLine(new PointF(Center.X + width / 2, (int)bounds.Top),
-											  new PointF(Center.X - width / 2, (int)bounds.Top),
-											  Location, out d4, out l4);
-			var minDistance = (new float[4] { d1, d2, d3, d4}).Min();
-			if (minDistance == d1)
-				return leftClosest;
-			else if (minDistance == d2)
-				return rightClosest;
-			else if (minDistance == d4)
-				return topClosest;
-			else
-				return bottomClosest;
-		}
-		
-		/// <summary>
-		/// Поиск ближайшей точки на линии относительно точки вне линии
-		/// </summary>
-		/// <param name="edgeStart">Позиция начала точки</param>
-		/// <param name="edgeEnd">Позиция конца точки</param>
-		/// <param name="p">точка относительно которой происходит поиск</param>
-		/// <param name="distance">Дистанция от точки до линии</param>
-		/// <returns>ближайшая точка</returns>
-		private PointF FindClosestPointOnLine(PointF edgeStart, PointF edgeEnd, PointF p, out float distance, out double lever)
-		{
-			float A = edgeStart.Y - edgeEnd.Y;
-			float B = edgeEnd.X - edgeStart.X;
-			float C = edgeStart.X * edgeEnd.Y - edgeEnd.X * edgeStart.Y;
-			float ab = A * A + B * B;
-			float cx = ((B * B * p.X - A * (B * p.Y + C)) / (float)ab);
-			float cy = (float)(A * ((-B) * p.X + A * p.Y) - B * C) / (float)ab;
-			lever = Math.Abs((A * (p.Y - edgeStart.Y) + B * (edgeStart.X - p.X)) / ab);
-			distance = (float)Math.Abs(A*p.X + B*p.Y + C) / (float)Math.Sqrt(ab);
-			if (cx > edgeStart.X || cx < edgeEnd.X)
-				cx = (edgeStart.X + edgeEnd.X) / 2;
-			if (cy > edgeStart.Y || cy < edgeEnd.Y)
-				cy = (edgeStart.Y + edgeEnd.Y) / 2;
-			return new PointF(cx, cy);
-		}
-	}
-
-	/// <summary>
-	/// Класс для визуального представления MArc
-	/// </summary>
-	[Serializable]
-	public class Line : IShape
-	{
-		/// <summary>
-		/// Определяет тип линии: обычная или inhibitor
-		/// </summary>
-		bool isInhibitor;
-		/// <summary>
-		/// Цвет рисования
-		/// </summary>
-		private Color DefaultColor = Color.Black;
-		/// <summary>
-		/// Цвет выделения
-		/// </summary>
-		private Color SelectedColor = Color.Green;
-		/// <summary>
-		/// Конструктор выставляющий по дефолту некоторые переменные
-		/// </summary>
-		public Line(bool isInhibitor)
-		{
-			LineWidth = 2; LineColor = Color.Black; points = new List<PointF>(); mArc = new MArc(isInhibitor);
-			this.isInhibitor = isInhibitor;
-		}
-		
-		/// <summary>
-		/// Шрифт для рисования некоторых подсказок на линии
-		/// </summary>
-		SerializableFont font = new SerializableFont("Arial", GraphicsUnit.Pixel, 12, FontStyle.Regular);
-		/// <summary>
-		/// Фигура к которой линия зацепилась в начале
-		/// </summary>
-		public IShape startShape { get; set; }
-		/// <summary>
-		/// Фигура к которой линия зацепилась в конце
-		/// </summary>
-		public IShape endShape { get; set; }
-		/// <summary>
-		/// Ширина рисуемой линии
-		/// </summary>
-		public int LineWidth { get; set; }
-		/// <summary>
-		/// Цвет линии
-		/// </summary>
-		public Color LineColor { get; set; }
-		/// <summary>
-		/// Масив точек для отрисовки кривой
-		/// </summary>
-		public List<PointF> points { get; set; }
-		/// <summary>
-		/// Соответсвующий объекту Line объект mArc 
-		/// </summary>
-		public MArc mArc;
-		/// <summary>
-		/// Соответсвующий объекту Line объект edge
-		/// </summary>
-		public Edge edge;
-		/// <summary>
-		/// Получить GraphicsPath для кривой
-		/// </summary>
-		/// <returns></returns>
-		public GraphicsPath GetPath()
-		{
-			var path = new GraphicsPath();
-			path.AddCurve(points.ToArray());
-			return path;
-		}
-		/// <summary>
-		/// Проверка попадания точки в GraphicsPath построенный по points
-		/// </summary>
-		/// <param name="p">Проверяемая точка</param>
-		/// <returns>Результат попадания</returns>
-		public bool HitTest(Point p)
-		{
-			var result = false;
-			using (var path = GetPath())
-			using (var pen = new Pen(LineColor, LineWidth + 8))
-				result = path.IsOutlineVisible(p, pen);
-			return result;
-		}
-		/// <summary>
-		/// Отрисовка кривой по точкам points
-		/// </summary>
-		/// <param name="g">место для отрисовки</param>
-		public void Draw(Graphics g)
-		{
-			var pen = new Pen(LineColor, LineWidth + 2);
-			if(!isInhibitor)
-				pen.EndCap = LineCap.ArrowAnchor;
-			if (isInhibitor)
-				pen.EndCap = LineCap.RoundAnchor;
-			using (var path = GetPath())
-			{
-				g.DrawPath(pen, path);
-				g.DrawString("In", font.ToFont(), Brushes.Brown, points[points.Count - 1].X - 40, points[points.Count - 1].Y - 20);
-				//Подсказска о весе линии
-				g.DrawString(String.Format("[{0}]", mArc.weight), font.ToFont(), Brushes.Brown, points[(points.Count - 1)/2].X, points[(points.Count - 1) / 2].Y + 20);
-			}
-		}
-		public void Move(Point d)
-		{
-		}
-		/// <summary>
-		/// Изменение размера линии путем замены последнего поинта в массиве точек points для кривой
-		/// </summary>
-		/// <param name="d">отклонение от последней точки</param>
-		public void Resize(Point d)
-		{
-			PointF newPoint = new PointF(points[points.Count - 1].X + d.X, points[points.Count - 1].Y + d.Y);
-			if (endShape == null)
-				points[points.Count - 1] = newPoint;
-			if (endShape != null)
-			{
-				var p = (endShape as INotArch).getClosestEdge(points[points.Count - 2]);
-				points[points.Count - 1] = new PointF(p.X, p.Y);
-			}
-		}
-		/// <summary>
-		/// Изменение размера линии путем замены первого поинта в массиве точек points для кривой
-		/// </summary>
-		/// <param name="d">отклонение от первой точки точки</param>
-		public void ResizeStart(Point d)
-		{
-			points[0] = new PointF(points[0].X + d.X, points[0].Y + d.Y);
-		}
-
-		public void RenameLabel(string newName) { }
-
-		/// <summary>
-		/// Изменение цвета и ширины для режима Select
-		/// </summary>
-		public void Select() {
-			LineColor = SelectedColor;
-			LineWidth = 4;
-		}
-		/// <summary>
-		/// Изменение цвета и ширины для режима Unselect
-		/// </summary>
-		public void Unselect()
-		{
-			LineColor = DefaultColor;
-			LineWidth = 2;
-		}
-		/// <summary>
-		/// Получить местоположение центра фигуры
-		/// </summary>
-		/// <returns>Point.Empty</returns>
-		public Point getCenter() { return Point.Empty; }
-
-		public PointF getEdge(Point curLocation) { return PointF.Empty; }
-
-		/// <summary>
-		/// Удалить объект Line со всеми его зависимостями из листа IShape
-		/// </summary>
-		/// <param name="shapes">лист откуда удаляем</param>
-		public bool delete(List<IShape> shapes)
-		{
-			try
-			{
-				Circle circle;
-				TRectangle rectangle;
-				List<MArc> lines;
-				if (this.startShape is Circle) { circle = (Circle)this.startShape; rectangle = (TRectangle)this.endShape; lines = ((TRectangle)this.endShape).model.inPlaces; }
-				else { circle = (Circle)this.endShape; rectangle = (TRectangle)this.startShape; lines = ((TRectangle)this.startShape).model.outPlaces; }
-				circle.inLines.Remove(this);
-				PetriNetsClassLibrary.PetriNet.CTransition.removeLink(this.mArc, lines);
-				rectangle.inLines.Remove(this);
-				shapes.Remove(this);
-				return true;
-			}
-			catch { return false; }
-		}
-
-
-	}
-
-	/// <summary>
-	/// Класс для изображения наименования объекта
-	/// </summary>
-	[Serializable]
-	public class Label
-	{
-		public Label(Point location, string text) { Location = location; Text = text;  }
-		public Point Location { get; set; }
-		public string Text { get; set; }
-		[NonSerialized]
-		private FontFamily FF;
-		public GraphicsPath GetPath()
-		{
-			var path = new GraphicsPath();
-			path.AddString(Text, FF = new FontFamily("Arial"), (int)FontStyle.Regular, 12, Location, StringFormat.GenericDefault);
-			return path;
-		}
-		public bool HitTest(Point p)
-		{
-			var result = false;
-			using (var path = GetPath())
-			using (var pen = new Pen(Color.Black, 4))
-				result = path.IsOutlineVisible(p, pen);
-			return result;
-		}
-		public void Draw(Graphics g)
-		{
-			using (var path = GetPath())
-			using (var pen = new Pen(Color.Black, 2))
-				g.DrawPath(pen, path);
-		}
-		public void Move(Point d)
-		{
-			Location = new Point(Location.X + d.X, Location.Y + d.Y);
-		}
-	}
-
-	[Serializable]
-	public class SerializableFont
-	{
-		public string FontFamily { get; set; }
-		public GraphicsUnit GraphicsUnit { get; set; }
-		public float Size { get; set; }
-		public FontStyle Style { get; set; }
-
-		/// <summary>
-		/// Intended for xml serialization purposes only
-		/// </summary>
-		private SerializableFont() { }
-
-		public SerializableFont(string fontFamily, GraphicsUnit graphicsUnit, float size, FontStyle style)
-		{
-			FontFamily = fontFamily;
-			GraphicsUnit = graphicsUnit;
-			Size = size;
-			Style = style;
-		}
-
-		public static SerializableFont FromFont(string fontFamily, GraphicsUnit graphicsUnit, float size, FontStyle style)
-		{
-			return new SerializableFont(fontFamily, graphicsUnit, size, style);
-		}
-
-		public Font ToFont()
-		{
-			return new Font(FontFamily, Size, Style, GraphicsUnit);
-		}
-	}
-	public interface IShape
-	{
-		GraphicsPath GetPath();
-		bool HitTest(Point p);
-		void Draw(Graphics g);
-		void Move(Point d);
-		void RenameLabel(string newName);
-		void Select();
-		void Unselect();
-		Point getCenter();
-		bool delete(List<IShape> shapes);
-	}
-
-	public interface INotArch
-	{
-		List<Line> getLines();
-		PointF getClosestEdge(PointF Location);
-	}
-
-	public abstract class Shape
-	{
-		public Color FillColor { get; set; }
-		public Label label { get; set; }
-	}
 }
