+++ PetriNetsClassLibrary/INotArch.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PetriNetsClassLibrary
+{
+    public interface INotArch
+    {
+        List<Line> getLines();
+        PointF getClosestEdge(PointF Location);
+    }
+}
