+++ PetriNetsClassLibrary/PetriNetsClassLibrary.csproj
@@ -52,8 +52,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Circle.cs" />
     <Compile Include="CTransition.cs" />
     <Compile Include="Edge.cs" />
+    <Compile Include="INotArch.cs" />
+    <Compile Include="IShape.cs" />
+    <Compile Include="Label.cs" />
+    <Compile Include="Line.cs" />
     <Compile Include="MArc.cs" />
     <Compile Include="Model.cs" />
     <Compile Include="MPlace.cs" />
@@ -65,6 +70,8 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <Compile Include="SerializableFont.cs" />
+    <Compile Include="TRectangle.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Properties\Resources.resx">
